pyproject.toml
[project]
name = "houseuk-astro"
version = "0.1.0"
description = "Data extraction and orchestration for UK housing datasets."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow-providers-cncf-kubernetes==10.11.1",
    "apache-airflow-providers-slack==9.5.0",
    "aws-lambda-powertools>=3.21.0,<4",
    "boto3==1.42.14",
    "fire==0.7.1",
    "kubernetes==34.1.0",
    "requests>=2.32.0,<3",
]

[dependency-groups]
dev = [
    "moto>=5.1.22",
    "mypy==1.18.2",
    "pre-commit==4.3.0",
    "pytest==9.0.0",
    "pytest-cov==7.0.0",
    "responses>=0.26.0",
    "ruff==0.13.2",
    "types-requests==2.32.4.20250913",
]
docs = [
    "mkdocs-material==9.7.3",
    "mkdocstrings==0.30.1",
    "mkdocstrings-python==2.0.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
target-version = "py313"
line-length = 120
extend-exclude = [".astro", "extract/crime/next/target"]

[tool.ruff.lint]
select = ["ALL"]


[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["T201"]
"examples/**/*.py" = ["T201"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_ignores = true
explicit_package_bases = true
files = ["extract", "lambda/stream_unzip_s3.py"]
